interfaces are refactored, connections are moved in services, repos are creating in services now

+++ server/src/main/java/ru/lavrov/tm/service/ProjectServiceImpl.java
@@ -5,7 +5,6 @@ import org.jetbrains.annotations.Nullable;
 import ru.lavrov.tm.api.IProjectRepository;
 import ru.lavrov.tm.api.IProjectService;
 import ru.lavrov.tm.api.ITaskRepository;
-import ru.lavrov.tm.api.IUserRepository;
 import ru.lavrov.tm.entity.Project;
 import ru.lavrov.tm.entity.Task;
 import ru.lavrov.tm.exception.general.DescriptionIsInvalidException;
@@ -14,25 +13,15 @@ import ru.lavrov.tm.exception.project.ProjectNameExistsException;
 import ru.lavrov.tm.exception.project.ProjectNameIsInvalidException;
 import ru.lavrov.tm.exception.project.ProjectNotExistsException;
 import ru.lavrov.tm.exception.user.UserIsNotAuthorizedException;
+import ru.lavrov.tm.repository.ProjectRepositoryImpl;
+import ru.lavrov.tm.repository.TaskRepositoryImpl;
 
+import java.nio.channels.ConnectionPendingException;
+import java.sql.Connection;
 import java.util.Collection;
 import java.util.Comparator;
 
-public final class ProjectServiceImpl extends AbstractService<Project> implements IProjectService {
-    @NotNull
-    protected final IProjectRepository projectRepository;
-    @NotNull
-    protected final ITaskRepository taskRepository;
-    @NotNull
-    protected final IUserRepository userRepository;
-
-    public ProjectServiceImpl(final IProjectRepository projectRepository, final ITaskRepository taskRepository,
-                              final IUserRepository userRepository) {
-        super(projectRepository);
-        this.projectRepository = projectRepository;
-        this.taskRepository = taskRepository;
-        this.userRepository = userRepository;
-    }
+public final class ProjectServiceImpl extends AbstractService implements IProjectService {
 
     @Override
     public void createByProjectName(@Nullable final String userId, @Nullable final String projectName) {
@@ -40,9 +29,13 @@ public final class ProjectServiceImpl extends AbstractService<Project> implement
             throw new ProjectNameIsInvalidException();
         if (userId == null || userId.isEmpty())
             throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
         if (projectRepository.findEntityByName(userId, projectName) != null)
             throw new ProjectNameExistsException();
-        persist(new Project(projectName, userId));
+        projectRepository.persist(new Project(projectName, userId));
     }
 
     @Override
@@ -51,10 +44,15 @@ public final class ProjectServiceImpl extends AbstractService<Project> implement
             throw new ProjectNameIsInvalidException();
         if (userId == null || userId.isEmpty())
             throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
+        @NotNull final ITaskRepository taskRepository = new TaskRepositoryImpl(connection);
         @Nullable final Project project = projectRepository.findEntityByName(userId, projectName);
         if (project != null)
             throw new ProjectNameExistsException();
-        removeProject(userId, project.getId());
+        projectRepository.removeProject(userId, project.getId());
         taskRepository.removeProjectTasks(userId, project.getId());
     }
 
@@ -64,7 +62,12 @@ public final class ProjectServiceImpl extends AbstractService<Project> implement
             throw new ProjectNameIsInvalidException();
         if (userId == null || userId.isEmpty())
             throw new UserIsNotAuthorizedException();
-        removeProject(userId, projectId);
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
+        @NotNull final ITaskRepository taskRepository = new TaskRepositoryImpl(connection);
+        projectRepository.removeProject(userId, projectId);
         taskRepository.removeProjectTasks(userId, projectId);
     }
 
@@ -75,6 +78,11 @@ public final class ProjectServiceImpl extends AbstractService<Project> implement
             throw new ProjectNameIsInvalidException();
         if (userId == null || userId.isEmpty())
             throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
+        @NotNull final ITaskRepository taskRepository = new TaskRepositoryImpl(connection);
         @Nullable final Project project = projectRepository.findEntityByName(userId, projectName);
         if (project == null)
             throw new ProjectNotExistsException();
@@ -86,12 +94,17 @@ public final class ProjectServiceImpl extends AbstractService<Project> implement
 
     @Nullable
     @Override
-    public void renameProject(@Nullable final String userId, @Nullable final String oldName,
-                              @Nullable final String newName) {
+    public void renameProject(
+            @Nullable final String userId, @Nullable final String oldName, @Nullable final String newName
+    )  {
         if (newName == null || newName.isEmpty() || oldName == null || oldName.isEmpty())
             throw new ProjectNameIsInvalidException();
         if (userId == null || userId.isEmpty())
             throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
         projectRepository.renameProject(userId, oldName, newName);
     }
 
@@ -102,6 +115,10 @@ public final class ProjectServiceImpl extends AbstractService<Project> implement
             throw new UserIsNotAuthorizedException();
         if (name == null || name.isEmpty())
             throw new NameIsInvalidException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
         @Nullable final Collection<Project> collection = projectRepository.findAllByNamePart(userId, name);
         return collection;
     }
@@ -113,6 +130,10 @@ public final class ProjectServiceImpl extends AbstractService<Project> implement
             throw new UserIsNotAuthorizedException();
         if (description == null || description.isEmpty())
             throw new DescriptionIsInvalidException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
         @Nullable final Collection<Project> collection = projectRepository.findAllByDescPart(userId, description);
         return collection;
     }
@@ -122,7 +143,56 @@ public final class ProjectServiceImpl extends AbstractService<Project> implement
     public Collection<Project> findAll(@Nullable final String userId, @Nullable final Comparator<Project> comparator) {
         if (userId == null || userId.isEmpty())
             throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
         return projectRepository.findAll(userId, comparator);
+    }
+
+    @Override
+    public void persist(@Nullable final Project entity) {
+        if (entity == null)
+            throw new ProjectNotExistsException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
+        projectRepository.persist(entity);
+    }
+
+    @Override
+    public void merge(@Nullable final Project entity) {
+        if (entity == null)
+            throw new ProjectNotExistsException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
+        projectRepository.persist(entity);
+    }
+
+    @Override
+    public void removeAll(@Nullable final String userId) {
+        if (userId == null || userId.isEmpty())
+            throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
+        projectRepository.removeAll(userId);
+    }
+
+    @Nullable
+    @Override
+    public Collection<Project> findAll(@Nullable final String userId) {
+        if (userId == null || userId.isEmpty())
+            throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IProjectRepository projectRepository = new ProjectRepositoryImpl(connection);
+        return projectRepository.findAll(userId);
     }
 }
