+++ server/src/main/java/ru/lavrov/tm/service/UserServiceImpl.java
@@ -9,25 +9,17 @@ import ru.lavrov.tm.api.IUserRepository;
 import ru.lavrov.tm.api.IUserService;
 import ru.lavrov.tm.entity.User;
 import ru.lavrov.tm.enumerate.Role;
+import ru.lavrov.tm.exception.project.ProjectNotExistsException;
 import ru.lavrov.tm.exception.user.*;
+import ru.lavrov.tm.repository.TaskRepositoryImpl;
+import ru.lavrov.tm.repository.UserRepositoryImpl;
 import ru.lavrov.tm.util.HashUtil;
 
-public final class UserServiceImpl extends AbstractService<User> implements IUserService {
-    @NotNull
-    protected final IProjectRepository projectRepository;
-    @NotNull
-    protected final ITaskRepository taskRepository;
-    @NotNull
-    protected final IUserRepository userRepository;
+import java.nio.channels.ConnectionPendingException;
+import java.sql.Connection;
+import java.util.Collection;
 
-    public UserServiceImpl(final IUserRepository userRepository,
-                           final IProjectRepository projectRepository,
-                           final ITaskRepository taskRepository) {
-        super(userRepository);
-        this.projectRepository = projectRepository;
-        this.taskRepository = taskRepository;
-        this.userRepository = userRepository;
-    }
+public final class UserServiceImpl extends AbstractService implements IUserService {
 
     public boolean createByLogin(
             @Nullable final String login, @Nullable final String password, @Nullable final String role
@@ -48,12 +40,16 @@ public final class UserServiceImpl extends AbstractService<User> implements IUse
             }
         if (!roleExists)
             throw new UserRoleIsInvalidException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
         @Nullable final User user = userRepository.findUserByLogin(login);
         if (user != null)
             throw new UserLoginExistsException();
         @NotNull String hashedPassword;
             hashedPassword = HashUtil.md5Hard(password);
-        persist(new User(login, hashedPassword, currentRole));
+        userRepository.persist(new User(login, hashedPassword, currentRole));
         return true;
     }
 
@@ -62,6 +58,10 @@ public final class UserServiceImpl extends AbstractService<User> implements IUse
             throw new UserIsNotAuthorizedException();
         if (newPassword == null || newPassword.isEmpty())
             throw new UserPasswordIsInvalidException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
         userRepository.updatePassword(userId, newPassword);
     }
 
@@ -70,6 +70,10 @@ public final class UserServiceImpl extends AbstractService<User> implements IUse
             throw new UserIsNotAuthorizedException();
         if (newLogin == null || newLogin.isEmpty())
             throw new UserLoginIsInvalidException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
         @Nullable final User user = userRepository.findUserByLogin(newLogin);
         if (user != null)
             throw new UserLoginExistsException();
@@ -81,6 +85,10 @@ public final class UserServiceImpl extends AbstractService<User> implements IUse
     public User findOne(@Nullable final String userId) {
         if (userId == null || userId.isEmpty())
             throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
         return userRepository.findOne(userId);
     }
 
@@ -89,6 +97,10 @@ public final class UserServiceImpl extends AbstractService<User> implements IUse
     public User findUserByLogin(@NotNull final String login) {
         if (login == null || login.isEmpty())
             throw new UserLoginIsInvalidException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
         return userRepository.findUserByLogin(login);
     }
 
@@ -96,6 +108,55 @@ public final class UserServiceImpl extends AbstractService<User> implements IUse
     public void removeUser(@Nullable final String userId) {
         if (userId == null || userId.isEmpty())
             throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
         userRepository.removeUser(userId);
+    }
+
+    @Override
+    public void persist(@Nullable final User entity) {
+        if (entity == null)
+            throw new ProjectNotExistsException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
+        userRepository.persist(entity);
+    }
+
+    @Override
+    public void merge(@Nullable final User entity) {
+        if (entity == null)
+            throw new ProjectNotExistsException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
+        userRepository.persist(entity);
+    }
+
+    @Override
+    public void removeAll(@Nullable final String userId) {
+        if (userId == null || userId.isEmpty())
+            throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
+        userRepository.removeAll(userId);
+    }
+
+    @Nullable
+    @Override
+    public Collection<User> findAll(@Nullable final String userId) {
+        if (userId == null || userId.isEmpty())
+            throw new UserIsNotAuthorizedException();
+        @Nullable final Connection connection = getConnection();
+        if (connection == null)
+            throw new ConnectionPendingException();
+        @NotNull final IUserRepository userRepository = new UserRepositoryImpl(connection);
+        return userRepository.findAll(userId);
     }
 }
